Cameras get their own crate
