+++ Cargo.toml
@@ -2,11 +2,13 @@
 members = [
     "libs/apriltag-sys",
     "libs/vv-apriltag",
+    "libs/vv-camera",
     "libs/vv-ntable",
+    "libs/vv-utils",
     "libs/vv-vision",
     "bins/vv-cli",
     "bins/vv-gui"
-, "libs/vv-utils"]
+]
 
 [workspace.package]
 version = "0.1.0"
@@ -21,29 +23,27 @@ license.workspace = true
 autoexamples = false
 
 [dependencies]
-disqualified = "1.0.0"
+disqualified.workspace = true
 humantime-serde = "1.1.1"
 litemap = "0.8.1"
 no_deadlocks = { version = "1.3.2", optional = true }
 papaya = "0.2.3"
-polonius-the-crab.workspace = true
 rayon.workspace = true
 serde.workspace = true
 smallvec.workspace = true
 smol_str = { workspace = true, features = ["serde"] }
-supply = "0.3.1"
+supply.workspace = true
 thiserror.workspace = true
 time.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 triomphe.workspace = true
-ty-tag = "0.1.1"
-typetag = "0.2.21"
-v4l = { workspace = true, optional = true }
+ty-tag.workspace = true
+typetag.workspace = true
 vv-apriltag = { workspace = true, optional = true, features = ["serde"] }
+vv-camera = { workspace = true, features = ["serde", "supply"] }
 vv-ntable = { workspace = true, optional = true }
 vv-vision = { workspace = true, default-features = false, features = ["serde"] }
-zune-jpeg = "0.5.8"
 
 [dev-dependencies]
 anyhow = "1.0"
@@ -53,11 +53,12 @@ tracing-subscriber = { workspace = true, features = ["env-filter"] }
 default = ["apriltag", "debug-gui", "ntable", "v4l"]
 apriltag = ["dep:vv-apriltag"]
 debug-gui = ["vv-vision/debug-gui"]
-v4l = ["dep:v4l", "vv-vision/v4l"]
+v4l = ["vv-camera/v4l", "vv-vision/v4l"]
 ntable = ["dep:vv-ntable"]
 debug-tools = ["dep:no_deadlocks"]
 
 [workspace.dependencies]
+disqualified = "1.0.0"
 eframe = { version = "0.33.2", features = ["persistence"] }
 polonius-the-crab = "0.5.0"
 rayon = "1.11.0"
@@ -65,15 +66,19 @@ serde = { version = "1.0", features = ["derive", "rc"] }
 serde_json = "1.0"
 smallvec = { version = "1.15.1", features = ["const_new", "drain_filter"] }
 smol_str = "0.3.4"
+supply = "0.3.1"
 thiserror = "2.0"
 time = { version = "0.3.44", features = ["formatting", "local-offset"] }
 toml = "0.9.8"
 tracing = "0.1.43"
 tracing-subscriber = "0.3.20"
 triomphe = "0.1.15"
+ty-tag = "0.1.1"
+typetag = "0.2.21"
 v4l = "0.14.0"
 viking-vision = { path = ".", default-features = false }
 vv-apriltag = { path = "./libs/vv-apriltag", default-features = false }
+vv-camera = { path = "./libs/vv-camera", default-features = false }
 vv-ntable = { path = "./libs/vv-ntable", default-features = false }
 vv-utils = { path = "./libs/vv-utils" }
 vv-vision = { path = "./libs/vv-vision", default-features = false }
