+++ libs/vv-utils/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod common_types;
 pub mod geom;
